requirements.txt
django-rest-passwordreset
django-filter
django-cors-headers
psycopg2
python-decouple
twilio